+++ test/reference_tests.jl
@@ -722,3 +722,58 @@ reftest("renamer latexstring rich text") do
     spec = data(df) * mapping(:x => rnm, :y, color = :x => rnm) * visual(Scatter)
     draw(spec)
 end
+
+function presorted_plot(; with_missing::Bool)
+    countries = ["Denmark", "Ecuador", "Bolivia", "China", "France", "Algeria"]
+    if with_missing
+        countries = [c == "Denmark" ? missing : c for c in countries]
+    end
+    group = ["2", "3", "1", "1", "3", "2"]
+    some_value = sort(exp.(sin.(1:6)))
+    
+    df = (; countries, group, some_value)
+
+    m1 = mapping(:countries, :some_value, color = :group)
+    m2 = mapping(:countries => presorted, :some_value, color = :group)
+    m3 = mapping(:countries => presorted, :some_value, color = :group => presorted)
+
+    base = data(df) *  visual(BarPlot, direction = :x)
+    
+    f = Figure()
+    fg1 = draw!(f[1, 1], base * m1)
+    fg2 = draw!(f[2, 1], base * m2)
+    fg3 = draw!(f[3, 1], base * m3)
+    legend!(f[1, 2], fg1)
+    legend!(f[2, 2], fg2)
+    legend!(f[3, 2], fg3)
+    f
+end
+
+reftest("presorted") do
+    presorted_plot(; with_missing = false)
+end
+
+reftest("presorted with missing") do
+    presorted_plot(; with_missing = true)
+end
+
+reftest("presorted additional data") do
+    df1 = (; name = ["Bob", "Charlie"], value = [27, 13])
+    df2 = (; name = ["Alice", "Bob"], value = [15, 30])
+
+    spec = data(df1) * visual(BarPlot)
+    spec2 = data(df2) * visual(Scatter)
+
+    f = Figure()
+    draw!(f[1, 1], (spec + spec2) * mapping(:name, :value))
+    draw!(f[1, 2], (spec + spec2) * mapping(:name => presorted, :value))
+
+    f
+end
+
+reftest("presorted wide") do
+    df = (; x1 = ["E", "D"], x2 = ["C", "B"], x3 = ["A", "E"], y = [1, 2])
+    spec = data(df) * mapping([:x1, :x2, :x3] .=> presorted, :y, color = dims(1)) *
+        visual(Scatter, markersize = 20)
+    draw(spec)
+end